+++ 04_components_v3/src/components/CompA.vue
@@ -1,0 +1,9 @@
+<script setup>
+import CompB from './CompB.vue';
+
+const { message } = defineProps(['message']);
+</script>
+<template>
+  <h2>CompAコンポーネント</h2>
+  <CompB :message="message" />
+</template>
