+++ 04_components_v3/src/main.js
@@ -7,6 +7,7 @@ import './style.css';
 // import App from './App/App05.vue';
 // import App from './App/App06.vue';
 // import App from './App/App07.vue';
+// import App from './App/App08.vue';
 import App from './App.vue';
 
 createApp(App).mount('#app');
